requirements.txt
aiogram==3.15.0
aiohttp==3.9.5
asyncpg==0.29.0
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.5.3